seeding: address issues observed during --reset
When running the installer with the --reset option, we have observed
errors.  This commit is intended to address the issues observed, so far. :)

First, the hooks provided by the fusor-installer should be executed after the
hooks of the base installer.

Second, it appears that although a --reset will remove the objects
from the database and backend services, the puppet environments
remain on the file system.  As a result, when the installer re-runs
it imports the environments in to the DB; however, it does not
properly associate the katello id, organization...etc.  Since
the installer will re-seed this environment, this commit will
simply remove the current environment.