seeding: disable assigning location to objects
Issues have been observed when locations are assigned to the
objects.  Since locations really are not a strict requirement
to support a deployment not going to provide that association
at this time.

One issue observed is that the install media that gets created
during a deployment is not getting assigned a location; however,
the host is.  As a result, if the user re-runs the installer
or even tries to update the location from the UI, they will be
denied the ability to do it, due to host existing in a location
but the media is not.